+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### New features
 - [PR #1121](https://github.com/rqlite/rqlite/pull/1121): Transparently forward node-removal requests to Leader.
 - [PR #1125](https://github.com/rqlite/rqlite/pull/1125): Support fetching a subset of expvar information.
+- [PR #1134](https://github.com/rqlite/rqlite/pull/1134): Support stepping down as Leader before shutting down.
 
 ### Implementation changes and bug fixes
 - [PR #1120](https://github.com/rqlite/rqlite/pull/1120): go mod updates.